Updates 30.03.2023 Part #1

+++ .pb-fixes/retropiemenu/Visuals/launchmp4.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 # Mp4 Launching Screens.
 # The PlayBox Project
-# Copyright (C)2018-2020 2Play! (S.R.)
-# 25.06.20
+# Copyright (C)2018-2023 2Play! (S.R.)
+# 25.06.2023
 
 infobox=""
 infobox="${infobox}\n"
@@ -48,17 +48,34 @@ function main_menu() {
 
 function enable_mp4() {
 	dialog --infobox "...Checking and Applying..." 3 31 ; sleep 2
+	#Check If ESBGM them apply correct files
+	if [ -f "/home/pi/.local/bin/esbgm" ]; then
+	cp /home/pi/PlayBox-Setup/mp4/runcommand-onstart2PESBGM.sh /opt/retropie/configs/all/runcommand-onstart.sh
+	else
 	cp /home/pi/PlayBox-Setup/mp4/runcommand-onstart2P.sh /opt/retropie/configs/all/runcommand-onstart.sh
+	fi
+	if [ ! -f "/usr/bin/loader" ]; then
 	sudo cp /usr/bin/omxplayer /usr/bin/loader
+	fi
 }
 
 function disable_mp4() {
 	dialog --infobox "...Checking and Applying..." 3 31 ; sleep 2
+	#Check If ESBGM them apply correct files
+	if [ -f "/home/pi/.local/bin/esbgm" ]; then
+	rm /opt/retropie/configs/all/runcommand-onstart.sh;
+	cp /home/pi/PlayBox-Setup/mp4/runcommand-onstartESBGM.sh /opt/retropie/configs/all/runcommand-onstart.sh
+	else
 	rm /opt/retropie/configs/all/runcommand-onstart.sh
+	fi
 }
 
 function enable_systems_mp4() {
 	dialog --infobox "...Checking if Systems mp4 exist..." 3 41 ; sleep 2
+	#Check If ESBGM them apply correct files
+	if [ -f "/home/pi/.local/bin/esbgm" ]; then
+	cp /home/pi/PlayBox-Setup/mp4/runcommand-onstartALLESBGM.sh /opt/retropie/configs/all/runcommand-onstart.sh
+	else
 	cp /home/pi/PlayBox-Setup/mp4/runcommand-onstartALL.sh /opt/retropie/configs/all/runcommand-onstart.sh
 }
